pyproject.toml
[tool.poetry]
name = "SuckYearBot"
version = "0.1.0"
description = "Лох года"
authors = ["MikhailTo <toshkin.mikhail@yandex.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
fluent-runtime = "^0.4.0"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
